feat: getParams 정적 함수로 추출
