+++ lib/index.ts
@@ -3,3 +3,4 @@ export { disableBrowserScrollRestoration } from "./components/BrowserRouter.help
 export { Link } from "./components/Link";
 export { matchDynamicRoute } from "./components/Router.helper";
 export { useRouter } from "./hooks/useRouter";
+export { getParams } from "./utils";
