+++ lib/utils.ts
@@ -13,3 +13,17 @@ export const make = (
     type,
   };
 };
+
+export const getParams = (history: History, path: string | undefined) => {
+  const pathParams = path
+    ?.split("/")
+    .map((path, index) => [path, index] as const)
+    .filter(([path]) => path.startsWith(":"));
+  const pathnames = history.pathname.split("/");
+  return {
+    ...history.query,
+    ...Object.fromEntries(
+      pathParams?.map(([path, index]) => [path, pathnames[index]]) ?? [],
+    ),
+  };
+};
